moved most code out of main
